Fixed security risk, improved site layout
A locked gem in Gemfile.lock had a security bug. Resolved this by removing Gemfile.lock from the repo, so only the latest GitHub Pages deps will be used.

Improved site layout and wording. Specifically reworked the projects and fun distractions pages to use a better description layout, and the home and education pages so that things I'm learning are all on the Education page.